Modified how app responses to the screen sizes
If screen is big, application shows 2 columns next to each other.
Trainings are located in the 1st one and exercises/sets in the 2nd.
If screen is small, after choosing the training user is transported
to the new view with exercises.

Also go back option was added in sign up view.